+++ lib/AudioOutputTest.ts
@@ -27,37 +27,37 @@ import {
   validateTime,
 } from './utils/optionValidation';
 
-export declare interface OutputTest {
+export declare interface AudioOutputTest {
   /**
    * This event is emitted when the test ends.
-   * @param event [[OutputTest.Events.End]]
+   * @param event [[AudioOutputTest.Events.End]]
    * @param report A summary of the test.
    * @private
    */
   emit(
-    event: OutputTest.Events.End,
-    report: OutputTest.Report,
+    event: AudioOutputTest.Events.End,
+    report: AudioOutputTest.Report,
   ): boolean;
   /**
    * This event is emitted when the test encounters an error, fatal or not.
-   * @param event [[OutputTest.Events.Error]]
+   * @param event [[AudioOutputTest.Events.Error]]
    * @param error An error that was encountered during the run time of the test.
    * @private
    */
   emit(
-    event: OutputTest.Events.Error,
+    event: AudioOutputTest.Events.Error,
     error: DiagnosticError,
   ): boolean;
   /**
    * This event is emitted by the test after succesfully starting, and emits
-   * the volume of the audio source every [[OutputTest.Options.volumeEventIntervalMs]]
+   * the volume of the audio source every [[AudioOutputTest.Options.volumeEventIntervalMs]]
    * milliseconds.
-   * @param event [[OutputTest.Events.Volume]]
+   * @param event [[AudioOutputTest.Events.Volume]]
    * @param value The volume of the audio source.
    * @private
    */
   emit(
-    event: OutputTest.Events.Volume,
+    event: AudioOutputTest.Events.Volume,
     value: number,
   ): boolean;
 
@@ -65,64 +65,64 @@ export declare interface OutputTest {
    * Raised when the test ends. The stop condition depends on if the option
    * to loop was set to `true` or `false`. If `false`, then the test ends either
    * when the audio file is finished playing, or when a time has elapsed
-   * greater than [[OutputTest.Options.duration]].
+   * greater than [[AudioOutputTest.Options.duration]].
    * @event
-   * @param event [[OutputTest.Events.End]]
+   * @param event [[AudioOutputTest.Events.End]]
    * @param listener A listener function that expects the following parameters
    * when the event is raised:
-   * - A [[OutputTest.Report]] that summarizes the run time of the test.
-   * @returns This [[OutputTest]] instance.
+   * - A [[AudioOutputTest.Report]] that summarizes the run time of the test.
+   * @returns This [[AudioOutputTest]] instance.
    */
   on(
-    event: OutputTest.Events.End,
-    listener: (report: OutputTest.Report) => any,
+    event: AudioOutputTest.Events.End,
+    listener: (report: AudioOutputTest.Report) => any,
   ): this;
   /**
    * Raised when the test has run into an error, fatal or not.
    * @event
-   * @param event [[OutputTest.Events.Error]]
+   * @param event [[AudioOutputTest.Events.Error]]
    * @param listener A listener function that expects the following parameters
    * when the event is raised:
    * - The [[DiagnosticError]].
-   * @returns This [[OutputTest]] instance.
+   * @returns This [[AudioOutputTest]] instance.
    */
   on(
-    event: OutputTest.Events.Error,
+    event: AudioOutputTest.Events.Error,
     listener: (error: DiagnosticError) => any,
   ): this;
   /**
-   * Raised every [[OutputTest.Options.volumeEventIntervalMs]] after the test
+   * Raised every [[AudioOutputTest.Options.volumeEventIntervalMs]] after the test
    * starts successfully. Will have a `number` parameter representing the
    * current volume of the audio file.
    * @event
-   * @param event [[OutputTest.Events.Volume]]
+   * @param event [[AudioOutputTest.Events.Volume]]
    * @param listener A listener function that expects the following parameters
    * when the event is raised:
    * - A number representing the volume of the audio source.
-   * @returns This [[OutputTest]] instance.
+   * @returns This [[AudioOutputTest]] instance.
    */
   on(
-    event: OutputTest.Events.Volume,
+    event: AudioOutputTest.Events.Volume,
     listener: (value: number) => any,
   ): this;
 }
 
 /**
- * [[OutputTest]] class that parses options and starts an audio output device
+ * [[AudioOutputTest]] class that parses options and starts an audio output device
  * test.
  *
- * Please see [[testOutputDevice]] for details and recommended practices.
+ * Please see [[testAudioOutputDevice]] for details and recommended practices.
  */
-export class OutputTest extends EventEmitter {
+export class AudioOutputTest extends EventEmitter {
   /**
    * The name of the test.
    */
   static testName: TestName.OutputAudioDevice = TestName.OutputAudioDevice;
   /**
-   * Default options for the [[OutputTest]]. Overwritten by any option passed
+   * Default options for the [[AudioOutputTest]]. Overwritten by any option passed
    * during the construction of the test.
    */
-  private static defaultOptions: OutputTest.InternalOptions = {
+  private static defaultOptions: AudioOutputTest.InternalOptions = {
     audioContextFactory: AudioContext,
     audioElementFactory: Audio,
     debug: false,
@@ -160,7 +160,7 @@ export class OutputTest extends EventEmitter {
    * Options passed to and set in the constructor to be used during the run
    * time of the test.
    */
-  private _options: OutputTest.InternalOptions;
+  private _options: AudioOutputTest.InternalOptions;
   /**
    * A timestamp of when the test starts. This is set in the constructor and not
    * when the test succesfully starts.
@@ -176,14 +176,14 @@ export class OutputTest extends EventEmitter {
   private _volumeTimeout: NodeJS.Timeout | null = null;
 
   /**
-   * Sets up several things for the `OutputTest` to run later in the
+   * Sets up several things for the [[AudioOutputTest]] to run later in the
    * `_startTest` function.
    * @param options Optional settings to pass to the test.
    */
-  constructor(options?: OutputTest.Options) {
+  constructor(options?: AudioOutputTest.Options) {
     super();
 
-    this._options = { ...OutputTest.defaultOptions, ...options };
+    this._options = { ...AudioOutputTest.defaultOptions, ...options };
 
     this._startTime = Date.now();
 
@@ -205,13 +205,13 @@ export class OutputTest extends EventEmitter {
     this._cleanup();
 
     this._endTime = Date.now();
-    const report: OutputTest.Report = {
+    const report: AudioOutputTest.Report = {
       deviceId: this._options.deviceId || (
         this._defaultDevices.audiooutput &&
         this._defaultDevices.audiooutput.deviceId
       ),
       errors: this._errors,
-      testName: OutputTest.testName,
+      testName: AudioOutputTest.testName,
       testTiming: {
         duration: this._endTime - this._startTime,
         end: this._endTime,
@@ -220,7 +220,7 @@ export class OutputTest extends EventEmitter {
       testURI: this._options.testURI,
       values: this._values,
     };
-    this.emit(OutputTest.Events.End, report);
+    this.emit(AudioOutputTest.Events.End, report);
   }
 
   /**
@@ -245,7 +245,7 @@ export class OutputTest extends EventEmitter {
    */
   private _onError(error: DiagnosticError): void {
     this._errors.push(error);
-    this.emit(OutputTest.Events.Error, error);
+    this.emit(AudioOutputTest.Events.Error, error);
   }
 
   /**
@@ -255,7 +255,7 @@ export class OutputTest extends EventEmitter {
    */
   private _onVolume(volume: number): void {
     this._values.push(volume);
-    this.emit(OutputTest.Events.Volume, volume);
+    this.emit(AudioOutputTest.Events.Volume, volume);
   }
 
   /**
@@ -273,15 +273,15 @@ export class OutputTest extends EventEmitter {
    * Entry point of the test, called after setup in the constructor.
    * Emits the volume levels of the audio.
    *
-   * @event [[OutputTest.Events.Volume]]
+   * @event [[AudioOutputTest.Events.Volume]]
    */
   private async _startTest(): Promise<void> {
     try {
       // Try to validate all of the inputs before starting the test.
       // We perform this check here so if the validation throws, it gets handled
       // properly as a fatal-error and we still emit a report with that error.
-      const invalidReasons: InvalidityRecord<OutputTest.Options> | undefined =
-        await validateOptions<OutputTest.Options>(this._options, {
+      const invalidReasons: InvalidityRecord<AudioOutputTest.Options> | undefined =
+        await validateOptions<AudioOutputTest.Options>(this._options, {
           deviceId: validateDeviceId,
           duration: validateTime,
           volumeEventIntervalMs: validateTime,
@@ -379,7 +379,7 @@ export class OutputTest extends EventEmitter {
       } else {
         if (this._options.deviceId && !setSinkIdSupported) {
           throw new UnsupportedError(
-            'A `deviceId` was passed to the `OutputTest` but `setSinkId` is ' +
+            'A `deviceId` was passed to the `AudioOutputTest` but `setSinkId` is ' +
             'not supported in this browser.',
           );
         }
@@ -422,10 +422,10 @@ export class OutputTest extends EventEmitter {
   }
 }
 
-export namespace OutputTest {
+export namespace AudioOutputTest {
   /**
-   * Events that the OutputTest will emit as it runs.
-   * Please see [[OutputTest.on]] for how to listen to these
+   * Events that the [[AudioOutputTest]] will emit as it runs.
+   * Please see [[AudioOutputTest.on]] for how to listen to these
    * events.
    */
   export enum Events {
@@ -435,7 +435,7 @@ export namespace OutputTest {
   }
 
   /**
-   * Options passed to [[OutputTest]] constructor.
+   * Options passed to [[AudioOutputTest]] constructor.
    */
   export interface Options {
     /**
@@ -465,7 +465,7 @@ export namespace OutputTest {
 
     /**
      * Whether or not to loop the audio.
-     * See [[OutputTest]] for details on the behavior of "timing out".
+     * See [[AudioOutputTest]] for details on the behavior of "timing out".
      * @default true
      */
     doLoop?: boolean;
@@ -473,7 +473,7 @@ export namespace OutputTest {
     /**
      * Duration in milliseconds to run the test for. If this amount of time elapses, the test
      * is considered "timed out".
-     * See [[OutputTest]] for details on the behavior of "timing out".
+     * See [[AudioOutputTest]] for details on the behavior of "timing out".
      * @default Infinity
      */
     duration?: number;
@@ -497,7 +497,7 @@ export namespace OutputTest {
   }
 
   /**
-   * Represents the report generated from an [[OutputTest]].
+   * Represents the report generated from an [[AudioOutputTest]].
    */
   export interface Report {
     /**
@@ -506,14 +506,14 @@ export namespace OutputTest {
     deviceId: string | undefined;
 
     /**
-     * Any errors that occurred during the run-time of the [[OutputTest]].
+     * Any errors that occurred during the run-time of the [[AudioOutputTest]].
      */
     errors: DiagnosticError[];
 
     /**
-     * Name of the test, set to [[OutputTest.testName]].
+     * Name of the test, set to [[AudioOutputTest.testName]].
      */
-    testName: typeof OutputTest.testName;
+    testName: typeof AudioOutputTest.testName;
 
     /**
      * Time measurements of test run time.
@@ -540,39 +540,39 @@ export namespace OutputTest {
 }
 
 /**
- * [[OutputTest]] tests audio output capabilities. It serves to help diagnose
+ * [[AudioOutputTest]] tests audio output capabilities. It serves to help diagnose
  * potential audio device issues that would prevent a user from being able to
  * hear audio.
  *
  * ---
  *
- * The [[OutputTest]] class is an `EventEmitter` (please see [[OutputTest.on]] for
+ * The [[AudioOutputTest]] class is an `EventEmitter` (please see [[AudioOutputTest.on]] for
  * events and their details) and helps to diagnose issues by playing a sound clip
  * (by default the sound clip is the ringing tone from the `twilio-client.js`
  * SDK) and emitting volume events of the sound clip as it plays.
  * ```ts
- * import { OutputTest, testOutputDevice } from '@twilio/rtc-diagnostics';
- * const options: OutputTest.Options = { ... };
+ * import { AudioOutputTest, testAudioOutputDevice } from '@twilio/rtc-diagnostics';
+ * const options: AudioOutputTest.Options = { ... };
  * // `options` may be left `undefined` to use default option values
- * const outputTest: OutputTest = testOutputDevice(options);
+ * const audioOutputTest: AudioOutputTest = testAudioOutputDevice(options);
  * ```
  * The application can use the volume events to show in its UI that audio is
  * playing and that the end-user should be hearing something.
  * ```ts
- * outputTest.on(OutputTest.Events.Volume, (volume: number) => {
+ * audioOutputTest.on(AudioOutputTest.Events.Volume, (volume: number) => {
  *   ui.updateVolume(volume); // Update your UI with the volume value here.
  * });
  * ```
  *
  * The application should ask the end-user to confirm that the sound being played
- * can be heard. The application should call [[OutputTest.stop]] with `true` if
+ * can be heard. The application should call [[AudioOutputTest.stop]] with `true` if
  * the end-user hears the sound, and `false` if not.
  * ```ts
  * // If the user was able to hear the audio, the UI should indicate they should
  * // click this button...
  * const passButton = ...;
  * passButton.on('click', () => {
- *   outputTest.stop();
+ *   audioOutputTest.stop();
  *   // display a confirmation dialog to the user
  * });
  *
@@ -580,23 +580,23 @@ export namespace OutputTest {
  * // this one.
  * const failButton = ...;
  * failButton.on('click', () => {
- *   outputTest.stop();
+ *   audioOutputTest.stop();
  *   // display a warning to the user
  * });
  * ```
- * Caling [[OutputTest.stop]] will immediately end the test.
+ * Caling [[AudioOutputTest.stop]] will immediately end the test.
  *
  * ---
  *
- * The [[OutputTest]] object will always emit a [[OutputTest.Report]] with
- * the [[OutputTest.Events.End]] event, regardless of the occurence of errors
+ * The [[AudioOutputTest]] object will always emit a [[AudioOutputTest.Report]] with
+ * the [[AudioOutputTest.Events.End]] event, regardless of the occurence of errors
  * during the runtime of the test.
  *
  * Fatal errors will immediately end the test and emit a report such that the
- * value of [[OutputTest.Report.errors]] will contain the fatal error.
+ * value of [[AudioOutputTest.Report.errors]] will contain the fatal error.
  *
  * Non-fatal errors will not end the test, but will be included in the value of
- * [[OutputTest.Report.errors]] upon completion of the test.
+ * [[AudioOutputTest.Report.errors]] upon completion of the test.
  *
  * If the data at `testURI` is unable to be loaded, meaning the error event is
  * raised on the audio element, a fatal error has occurred.
@@ -608,18 +608,18 @@ export namespace OutputTest {
  *
  * ---
  *
- * The function [[testOutputDevice]] serves as factory function that accepts
- * [[OutputTest.Options]] as its only parameter and will instantiate an
- * [[OutputTest]] object with those options.
+ * The function [[testAudioOutputDevice]] serves as factory function that accepts
+ * [[AudioOutputTest.Options]] as its only parameter and will instantiate an
+ * [[AudioOutputTest]] object with those options.
  * ```ts
- * import { OutputTest, testOutputDevice } from '@twilio/rtc-diagnostics';
- * const options: OutputTest.Options = { ... };
- * const outputTest: OutputTest = testOutputDevice(options);
+ * import { AudioOutputTest, testAudioOutputDevice } from '@twilio/rtc-diagnostics';
+ * const options: AudioOutputTest.Options = { ... };
+ * const audioOutputTest: AudioOutputTest = testAudioOutputDevice(options);
  * ```
- * @param options Options to pass to the [[OutputTest]] constructor.
+ * @param options Options to pass to the [[AudioOutputTest]] constructor.
  */
-export function testOutputDevice(
-  options?: OutputTest.Options,
-): OutputTest {
-  return new OutputTest(options);
+export function testAudioOutputDevice(
+  options?: AudioOutputTest.Options,
+): AudioOutputTest {
+  return new AudioOutputTest(options);
 }